requirements.txt
graphene[django]
django_graphiql
djangorestframework
django-rest-swagger
graphql-core
django==1.9
django-filter==0.11.0

